# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Maids.ng API
  description: |
    # Maids.ng API - Agentic Era Ready
    
    A comprehensive REST API for domestic help matching and management, designed for the "Agentic Era" with AI-friendly structured responses.
    
    ## Key Features
    - **Standardized Response Format**: All responses follow a consistent envelope structure
    - **AI-Optimized**: Clear schemas, metadata, and request IDs for agent consumption
    - **Role-Based Access**: Separate endpoints for Employers, Maids, and Admins
    - **Comprehensive Coverage**: Authentication, Profiles, Bookings, Payments, and Admin operations
    
    ## Authentication
    This API uses Laravel Sanctum for token-based authentication. 
    
    1. Obtain a token via `POST /api/v1/auth/login`
    2. Include the token in the Authorization header: `Authorization: Bearer {token}`
    
    ## Response Format
    All responses follow this structure:
    ```json
    {
      "success": true,
      "message": "Operation completed",
      "data": { ... },
      "meta": {
        "timestamp": "2024-01-15T10:30:00Z",
        "request_id": "uuid",
        "api_version": "1.0.0"
      }
    }
    ```
    
    ## Rate Limiting
    - Public endpoints: 60 requests per minute
    - Authenticated endpoints: 1000 requests per minute
    
  version: 1.0.0
  contact:
    name: Maids.ng API Support
    email: api@maids.ng
  license:
    name: Proprietary
    url: https://maids.ng/terms
servers:
...
security:
  - bearerAuth: []
paths:
...
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register new user
      description: Create a new employer or maid account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '422':
          description: Validation error
...
  /maid/profile:
    get:
      tags:
        - Maid Profile
      summary: Get my profile
      description: Retrieve authenticated maid's profile
      responses:
        '200':
          description: Profile retrieved
    put:
      tags:
        - Maid Profile
      summary: Update profile
      description: Update maid profile information
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMaidProfileRequest'
      responses:
        '200':
          description: Profile updated
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RegisterRequest:
      type: object
      required:
        - name
        - email
        - phone
        - password
        - password_confirmation
        - role
      properties:
        name:
          type: string
          maxLength: 255
        email:
          type: string
          format: email
        phone:
          type: string
          maxLength: 20
        password:
          type: string
          minLength: 8
        password_confirmation:
          type: string
        role:
          type: string
          enum: [employer, maid]
        location:
          type: string
...
    UpdateMaidProfileRequest:
      type: object
      properties:
        bio:
          type: string
          maxLength: 1000
        skills:
          type: array
          items:
            type: string
        experience_years:
          type: integer
          minimum: 0
          maximum: 50
        help_types:
          type: array
          items:
            type: string
            enum: [live-in, nanny, cooking, elderly-care, driver, cleaning, laundry]
        schedule_preference:
          type: string
          enum: [full-time, part-time, weekends, flexible]
        expected_salary:
          type: integer
          minimum: 10000
          maximum: 500000
        location:
          type: string
        state:
          type: string
        lga:
          type: string
        availability_status:
          type: string
          enum: [available, busy, unavailable]
...
    AuthResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/User'
            token:
              type: string
            token_type:
              type: string
        meta:
          $ref: '#/components/schemas/Meta'
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        avatar:
          type: string
        status:
          type: string
        location:
          type: string
        roles:
          type: array
          items:
            type: string
        maid_profile:
          $ref: '#/components/schemas/MaidProfile'
        employer_preferences:
          type: array
          items:
            $ref: '#/components/schemas/EmployerPreference'
        created_at:
          type: string
          format: date-time
    MaidProfile:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        bio:
          type: string
        role:
          type: string
        skills:
          type: array
          items:
            type: string
        work_type:
          type: string
        schedule_preference:
          type: string
        availability_status:
          type: string
        location:
          type: string
        city:
          type: string
        state:
          type: string
        expected_salary:
          type: integer
        salary_currency:
          type: string
        experience_years:
          type: integer
        rating:
          type: number
        total_reviews:
          type: integer
        total_bookings:
          type: integer
        verification:
          type: object
          properties:
            nin_verified:
              type: boolean
            background_verified:
              type: boolean
            fully_verified:
              type: boolean
        user:
          type: object
          properties:
            id:
              type: integer
            name:
              type: string
            avatar:
              type: string
    EmployerPreference:
      type: object
      properties:
        id:
          type: integer
        employer_id:
          type: integer
        help_type:
          type: string
        location:
          type: string
        state:
          type: string
        schedule_type:
          type: string
        salary_budget:
          type: integer
        required_skills:
          type: array
          items:
            type: string
        status:
          type: string
    Meta:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        request_id:
          type: string
        api_version:
          type: string
...